requirements.txt
python-dotenv
pika
psycopg2-binary
google-generativeai
pgvector